+++ dymension/libs/kaspa/demo/relayer/src/x/deposit.rs
@@ -45,56 +45,6 @@ use api_rs::apis::kaspa_transactions_api::{
 use kaspa_rpc_core::api::rpc::RpcApi;
 use workflow_core::abortable::Abortable;
 
-pub async fn deposit(
-    w: &Arc<Wallet>,
-    secret: &Secret,
-    address: Address,
-    amt: u64,
-) -> Result<TransactionId, KaspaError> {
-    let mut hl_message = HyperlaneMessage::default();
-    let token_message = TokenMessage::new(H256::random(), U256::from(amt), vec![]);
-
-    let encoded_bytes = token_message.to_vec();
-
-    hl_message.body = encoded_bytes;
-
-    let payload = hl_message.to_vec();
-
-    deposit_impl(w, secret, address.clone(), amt, payload.clone()).await
-}
-
-pub async fn deposit_impl(
-    w: &Arc<Wallet>,
-    secret: &Secret,
-    address: Address,
-    amt: u64,
-    payload: Vec<u8>,
-) -> Result<TransactionId, KaspaError> {
-    let a = w.account()?;
-
-    let dst = PaymentDestination::from(PaymentOutput::new(address, amt));
-    let fees = Fees::from(0i64);
-    let payment_secret = None;
-    let abortable = Abortable::new();
-
-    // use account.send, because wallet.accounts_send(AccountsSendRequest{..}) is buggy
-    let (summary, _) = a
-        .send(
-            dst,
-            fees,
-            Some(payload),
-            secret.clone(),
-            payment_secret,
-            &abortable,
-            None,
-        )
-        .await?;
-
-    summary.final_transaction_id().ok_or_else(|| {
-        KaspaError::Custom("Deposit transaction failed to generate a transaction ID".to_string())
-    })
-}
-
 pub struct DemoArgs {
     pub amt: u64,
     pub escrow_address: Address,
