+++ dymension/libs/kaspa/lib/core/Cargo.toml
@@ -4,6 +4,8 @@ version = "0.2.0"
 edition = "2021"
 
 [dependencies]
+hardcode = { path = "../hardcode" }
+hex = "0.4"
 kaspa-addresses = { workspace = true }
 kaspa-consensus-core = { workspace = true }
 kaspa-grpc-client = { workspace = true }
@@ -37,5 +39,4 @@ serde-value = "0.7.0"
 reqwest-middleware = "0.4"
 reqwest-ratelimit = "0.4.1"
 governor = "0.10"
-anyhow = "1.0"
 serde = { version = "^1.0", features = ["derive"] }
