+++ dymension/libs/kaspa/lib/core/src/user/mod.rs
@@ -1,0 +1,1 @@
+mod deposit;
